pyproject.toml
[tool.ruff.lint]
ignore = ["E501"]

[tool.pytest.ini_options]
# MUST INSTALL pip install pytest-env for neotest-python to work
env = ["DJANGO_SETTINGS_MODULE=project.settings"]

[project]
name = "project"
version = "0.1.0"
description = ""
authors = [{ name = "strongmove@strongmove.home" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
  "django (>=5.2,<6.0)",
  "daphne (>=4.1.2,<5.0.0)",
  "django-compressor (>=4.5.1,<5.0.0)",
  "django-allauth (>=65.7.0,<66.0.0)",
  "django-cors-headers (>=4.7.0,<5.0.0)",
  "django-htmx (>=1.23.0,<2.0.0)",
  "django-browser-reload (>=1.18.0,<2.0.0)",
  "psycopg2-binary (>=2.9.10,<3.0.0)",
  "requests (>=2.32.3,<3.0.0)",
  "pyjwt (>=2.10.1,<3.0.0)",
  "whitenoise (>=6.9.0,<7.0.0)",
  "django-stubs (>=5.2.0,<6.0.0)",
  "strawberry-graphql-django (>=0.59.0,<0.60.0)",
  "django-choices-field (>=2.3.0,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
